+++ src/Domain/Infra/DependencyInjection/ExtensionHelper.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace MsgPhp\Domain\Infra\DependencyInjection;
 
+use Doctrine\DBAL\Types\Type as DoctrineType;
+use Ramsey\Uuid\Doctrine as DoctrineUuid;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 
 /**
@@ -13,37 +15,78 @@ use Symfony\Component\DependencyInjection\ContainerBuilder;
  */
 final class ExtensionHelper
 {
-    public static function configureIdentityMapping(ContainerBuilder $container, array $classMapping, array $identityMapping): void
-    {
-        foreach ($identityMapping as $class => $mapping) {
-            if (isset($classMapping[$class]) && !isset($identityMapping[$classMapping[$class]])) {
-                $identityMapping[$classMapping[$class]] = $mapping;
-            }
-        }
-
-        $values = $container->hasParameter($param = 'msgphp.domain.identity_mapping') ? $container->getParameter($param) : [];
-        $values[] = $identityMapping;
-
-        $container->setParameter($param, $values);
-    }
-
-    public static function configureEntityFactory(ContainerBuilder $container, array $classMapping, array $idClassMapping): void
+    public static function configureDomain(ContainerBuilder $container, array $classMapping, array $idClassMapping, array $identityMapping): void
     {
         foreach ($idClassMapping as $class => $idClass) {
             if (isset($classMapping[$class]) && !isset($idClassMapping[$classMapping[$class]])) {
                 $idClassMapping[$classMapping[$class]] = $idClass;
             }
         }
+        foreach ($identityMapping as $class => $mapping) {
+            if (isset($classMapping[$class]) && !isset($identityMapping[$classMapping[$class]])) {
+                $identityMapping[$classMapping[$class]] = $mapping;
+            }
+        }
 
         $values = $container->hasParameter($param = 'msgphp.domain.class_mapping') ? $container->getParameter($param) : [];
         $values[] = $classMapping;
-
         $container->setParameter($param, $values);
 
         $values = $container->hasParameter($param = 'msgphp.domain.id_class_mapping') ? $container->getParameter($param) : [];
         $values[] = $idClassMapping;
-
         $container->setParameter($param, $values);
+
+        $values = $container->hasParameter($param = 'msgphp.domain.identity_mapping') ? $container->getParameter($param) : [];
+        $values[] = $identityMapping;
+        $container->setParameter($param, $values);
+    }
+
+    public static function configureDoctrineOrm(ContainerBuilder $container, array $classMapping, array $idTypeMapping, array $typeClassMapping): void
+    {
+        $dbalTypes = $dbalMappingTypes = $typeConfig = [];
+        $uuidMapping = [
+            'uuid' => DoctrineUuid\UuidType::class,
+            'uuid_binary' => DoctrineUuid\UuidBinaryType::class,
+            'uuid_binary_ordered_time' => DoctrineUuid\UuidBinaryOrderedTimeType::class,
+        ];
+
+        foreach ($typeClassMapping as $idClass => $typeClass) {
+            $type = $idTypeMapping[$idClass] ?? DoctrineType::INTEGER;
+
+            if (isset($uuidMapping[$type])) {
+                if (!class_exists($uuidClass = $uuidMapping[$type])) {
+                    throw new \LogicException(sprintf('Type "%s" for identifier "%s" requires "ramsey/uuid-doctrine".', $type, $idClass));
+                }
+
+                $dbalTypes[$uuidClass::NAME] = $uuidClass;
+
+                if ('uuid_binary' === $type || 'uuid_binary_ordered_time' === $type) {
+                    $dbalMappingTypes[$type] = 'binary';
+                }
+            }
+
+            if (!defined($typeClass.'::NAME')) {
+                throw new \LogicException(sprintf('Type class "%s" for identifier "%s" requires a "NAME" constant.', $typeClass, $idClass));
+            }
+
+            $dbalTypes[$typeClass::NAME] = $typeClass;
+            $typeConfig[$typeClass::NAME] = ['class' => $classMapping[$idClass] ?? $idClass, 'type' => $type, 'type_class' => $typeClass];
+        }
+
+        if ($container->hasParameter($param = 'msgphp.doctrine.type_config')) {
+            $typeConfig += $container->getParameter($param);
+        }
+        $container->setParameter($param, $typeConfig);
+
+        $container->prependExtensionConfig('doctrine', [
+            'dbal' => [
+                'types' => $dbalTypes,
+                'mapping_types' => $dbalMappingTypes,
+            ],
+            'orm' => [
+                'resolve_target_entities' => $classMapping,
+            ],
+        ]);
     }
 
     private function __construct()
